Add files via upload
I created an interface to automate the task of creating newsletters that are used by many colleges to keep the students posted about the campus events. This interface asks for information about the event from the admin and creates a html webpage which can be embedded into emails and can be sent out to students.